+++ src/util/googleMaps.js
@@ -36,17 +36,21 @@ const placeCountry = place => {
  *
  * @param {String} placeId - ID for a place received from the
  * autocomplete service
+ * @param {String} sessionToken - token to tie different autocomplete character searches together
+ * with getPlaceDetails call
  *
  * @return {Promise<util.propTypes.place>} Promise that
  * resolves to the detailed place, rejects if the request failed
  */
-export const getPlaceDetails = placeId =>
+export const getPlaceDetails = (placeId, sessionToken) =>
   new Promise((resolve, reject) => {
     const serviceStatus = window.google.maps.places.PlacesServiceStatus;
     const el = document.createElement('div');
     const service = new window.google.maps.places.PlacesService(el);
+    const fields = ['address_component', 'formatted_address', 'geometry', 'place_id'];
+    const sessionTokenMaybe = sessionToken ? { sessionToken } : {};
 
-    service.getDetails({ placeId }, (place, status) => {
+    service.getDetails({ placeId, fields, ...sessionTokenMaybe }, (place, status) => {
       if (status !== serviceStatus.OK) {
         reject(
           new Error(`Could not get details for place id "${placeId}", error status was "${status}"`)
@@ -71,16 +75,19 @@ const predictionSuccessful = status => {
  * Get place predictions for the given search
  *
  * @param {String} search - place name or address to search
+ * @param {String} sessionToken - token to tie different autocomplete character searches together
+ * with getPlaceDetails call
  *
  * @return {Promise<{ search, predictions[] }>} - Promise of an object
  * with the original search query and an array of
  * `google.maps.places.AutocompletePrediction` objects
  */
-export const getPlacePredictions = search =>
+export const getPlacePredictions = (search, sessionToken) =>
   new Promise((resolve, reject) => {
     const service = new window.google.maps.places.AutocompleteService();
+    const sessionTokenMaybe = sessionToken ? { sessionToken } : {};
 
-    service.getPlacePredictions({ input: search }, (predictions, status) => {
+    service.getPlacePredictions({ input: search, ...sessionTokenMaybe }, (predictions, status) => {
       if (!predictionSuccessful(status)) {
         reject(new Error(`Prediction service status not OK: ${status}`));
       } else {