+++ CHANGELOG.md
@@ -13,7 +13,9 @@ way to update this template, but currently, we follow a pattern:
 ---
 
 ## Upcoming version
-
+* [change] Use sessionTokens and fields for Autocomplete calls to Google Maps.
+  This is a reaction to pricing change of Google Maps APIs.
+  [#867](https://github.com/sharetribe/flex-template-web/pull/867)
 * [change] Change TransactionPage state management in loadData.
   [#863](https://github.com/sharetribe/flex-template-web/pull/863), [#865](https://github.com/sharetribe/flex-template-web/pull/865) & [#866](https://github.com/sharetribe/flex-template-web/pull/866)
 * [fix] Fix submit button state on contact details page.